+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.17.44
+### is.codion.swing.framework.ui
+- EntityEditPanel, control configuration refactored, now consistent with EntityTablePanel.
 
 ## 0.17.43
 ### is.codion.common.core